Move search radius to Near line in voting form
Show the search radius bubble next to "Near …" at the top of
location/restaurant polls instead of as a separate "Search within"
select above the suggestion input. Tapping the bubble opens the same
modal used in the poll-creation form.

Extracted the bubble + modal as SearchRadiusBubble so the voting form
and the creation form share one implementation.
